requirements.txt
# SquadOps Agent Runtime Dependencies
# SIP-0.8.8: New Architecture

# Core frameworks
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.7.3
jinja2>=3.1.2

# Async utilities
aiohttp==3.9.1
aiofiles==23.2.1

# Message queue
aio-pika==9.3.1
pika>=1.3.2

# Database
asyncpg==0.29.0
redis==5.0.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# System utilities
psutil==5.9.8

# Memory system (SIP-042)
lancedb==0.8.2
sentence-transformers>=2.7.0
pyarrow>=14.0.0
pandas>=2.0.0

# OpenTelemetry
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-exporter-prometheus
opentelemetry-semantic-conventions

# LLM Observability (SIP-0061)
langfuse>=2.0,<3.0

# Auth (SIP-0062)
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Validation
jsonschema>=4.20.0

# State management
transitions>=0.9.0
